[refactor] Add base class GfxProgramImpl
[ghstack-poisoned]
