+++ taichi/runtime/program_impls/dx/dx_program.cpp
@@ -13,7 +13,7 @@
 
 namespace taichi::lang {
 
-Dx11ProgramImpl::Dx11ProgramImpl(CompileConfig &config) : ProgramImpl(config) {
+Dx11ProgramImpl::Dx11ProgramImpl(CompileConfig &config) : GfxProgramImpl(config) {
 }
 
 void Dx11ProgramImpl::materialize_runtime(KernelProfilerBase *profiler,
@@ -27,64 +27,6 @@ void Dx11ProgramImpl::materialize_runtime(KernelProfilerBase *profiler,
   params.device = device_.get();
   runtime_ = std::make_unique<gfx::GfxRuntime>(std::move(params));
   snode_tree_mgr_ = std::make_unique<gfx::SNodeTreeManager>(runtime_.get());
-}
-
-void Dx11ProgramImpl::compile_snode_tree_types(SNodeTree *tree) {
-  if (runtime_) {
-    snode_tree_mgr_->materialize_snode_tree(tree);
-  } else {
-    gfx::CompiledSNodeStructs compiled_structs =
-        gfx::compile_snode_structs(*tree->root());
-    aot_compiled_snode_structs_.push_back(compiled_structs);
-  }
-}
-
-void Dx11ProgramImpl::materialize_snode_tree(SNodeTree *tree,
-                                             uint64 *result_buffer) {
-  snode_tree_mgr_->materialize_snode_tree(tree);
-}
-
-std::unique_ptr<AotModuleBuilder> Dx11ProgramImpl::make_aot_module_builder(
-    const DeviceCapabilityConfig &caps) {
-  if (runtime_) {
-    return std::make_unique<gfx::AotModuleBuilderImpl>(
-        snode_tree_mgr_->get_compiled_structs(),
-        get_kernel_compilation_manager(), *config, caps);
-  } else {
-    return std::make_unique<gfx::AotModuleBuilderImpl>(
-        aot_compiled_snode_structs_, get_kernel_compilation_manager(), *config,
-        caps);
-  }
-}
-
-DeviceAllocation Dx11ProgramImpl::allocate_memory_ndarray(
-    std::size_t alloc_size,
-    uint64 *result_buffer) {
-  DeviceAllocation alloc;
-  RhiResult res = get_compute_device()->allocate_memory(
-      {alloc_size, /*host_write=*/false, /*host_read=*/false,
-       /*export_sharing=*/false},
-      &alloc);
-  TI_ASSERT(res == RhiResult::success);
-  return alloc;
-}
-
-std::unique_ptr<KernelCompiler> Dx11ProgramImpl::make_kernel_compiler() {
-  spirv::KernelCompiler::Config cfg;
-  cfg.compiled_struct_data = runtime_ ? &snode_tree_mgr_->get_compiled_structs()
-                                      : &aot_compiled_snode_structs_;
-  return std::make_unique<spirv::KernelCompiler>(std::move(cfg));
-}
-
-std::unique_ptr<KernelLauncher> Dx11ProgramImpl::make_kernel_launcher() {
-  gfx::KernelLauncher::Config cfg;
-  cfg.gfx_runtime_ = runtime_.get();
-  return std::make_unique<gfx::KernelLauncher>(std::move(cfg));
-}
-
-DeviceCapabilityConfig Dx11ProgramImpl::get_device_caps() {
-  TI_ASSERT(runtime_);
-  return runtime_->get_ti_device()->get_caps();
 }
 
 }  // namespace taichi::lang
