+++ taichi/runtime/program_impls/gfx/CMakeLists.txt
@@ -1,0 +1,19 @@
+# ./taichi/runtime/program_impls/gfx/CMakeLists.txt
+
+add_library(gfx_program_impl)
+target_sources(gfx_program_impl
+        PRIVATE
+        gfx_program.cpp
+        )
+
+target_include_directories(gfx_program_impl
+        PRIVATE
+        ${PROJECT_SOURCE_DIR}
+        ${PROJECT_SOURCE_DIR}/external/eigen
+        ${PROJECT_SOURCE_DIR}/external/spdlog/include
+        ${PROJECT_SOURCE_DIR}/external/SPIRV-Tools/include
+        ${LLVM_INCLUDE_DIRS}
+        )
+
+target_link_libraries(gfx_program_impl PRIVATE gfx_runtime)
+target_link_libraries(gfx_program_impl PRIVATE compilation_manager)
