+++ taichi/runtime/program_impls/vulkan/CMakeLists.txt
@@ -19,6 +19,7 @@ target_include_directories(vulkan_program_impl
 
 target_link_libraries(vulkan_program_impl PRIVATE spirv_codegen)
 target_link_libraries(vulkan_program_impl PRIVATE gfx_runtime)
+target_link_libraries(opengl_program_impl PRIVATE gfx_program_impl)
 target_link_libraries(vulkan_program_impl PRIVATE compilation_manager)
 target_link_libraries(vulkan_program_impl PRIVATE ti_device_api)
 target_link_libraries(vulkan_program_impl PRIVATE vulkan_rhi)
